Add support for db locking in zsh (via flock)
This commit ensures the database is locked before reading and updating
when used in zsh. We load the zsh/system module which provides the
zsystem flock command to perform advisory file locking.

A separate code path is needed when updating the database entry because
replacing the file (via mv) would cause any process waiting for a lock
to fail. Here we simply clobber the database with the contents of the
tempfile.